Add `InterfaceElement` ABC.
Change many of the classes to derive from the InterfaceElement ABC. This class
is used to generally define the relationship between an instance and the
"creator", or more appropriately, ancestor. This will lead to more informative
exceptions, warnings, and messages.

Eventually this model will be generalized outside of py-postgresql, but for now,
just define the primary components here.