Add the remaining tests to testall.
Conditionally include test_ssl_connect based on the installation's SSL support.

Perhaps this should be left to fail, but currently the initialization of the
testcase depends on the postmaster being able to startup. pg will refuse to start
due to the lack of compiled-in SSL support and an exception will be raised causing
the remaining tests to not even be ran. Just skip it for now.