Remove hard parameters and the -1 option.
hard parameters were basically unused, and the effect of -1 is not so useful as
a runtime option in python scripts.(well, i suppose it could be, but errors
would have to be treated as messages)

Also, prompt the user for a new password on authspec connection failures.
However, this should probably be generalized to allow the user to reconfigure
the connection parameters entirely.