Fix an invalid state bug on cursors.
This issue was caused by an invalid quantity being expressed in specific
read() invocations. The quantity was expressed as the utter remainder that
needed to be read, where it needed to be the the total number read. The issue
than manifested itself as providing an offset greater than the length of the
current buffer. This caused an over-compensation on future reads due to an
effective negative available rows.

Unrelated cleanups being some additional whitespace around statement snapshots.