Implement TupleError for providing details on pack/unpack failures.
This gives the user information about what failed when a row could not be packed
or unpacked. It uses exception __context__'s to reference the original failure
while providing the user with an appropriately general error leading them to the
specific piece of data in the tuple that was invalid.

Some additional cleanups on the error's details may be necessary. It's probably
using too many fields, atm. Additionally, "NUMBER" and "NAME" labels are too
ambiguous.