Relocate the chunks() iterator.
Based on the assumed use-case of chunks(), it was relocated to statement
objects.

The likely fact is that chunks would be solely used from the beginning and to
the end of a cursor. It should not be used with scrollable cursors and it
expects receiving a fair amount of data so WITH HOLD would always be true
outside of transactions.

Additionally, give it a chunksize keyword argument.