Change the default invocation of statements to return the entire result-set.
The default mechanism should probably be the most simple, and a cursor is
certainly not the most simple. Additionally, calling a DML statement will return
a tuple: (cmd, count).

With this comes two new execution methods:

 declare() : create a formal, scrollable cursor with hold for random access.
 rows() : wrapper around chunks() for processing large result-sets.

This should cover most of the possible execution methods that would be desired
by a user.

[A big plus here is mostly ridding the higher level interfaces of keyword arguments]