Refactor InterfaceElement in order to provide "multiple ancestors".
The InterfaceElement ABC has be moved to postgresql.python.element.Element.
"Ancestors" are no more. Rather, they are called factors. As
in, the factors that play a critical role in the definition of this element.

ife_snapshot_text was completely removed in favor of
postgresql.python.element.format_element and the _e_metas methods.
format_element takes care of all the formatting cases for the printing of the
factor tree, and _e_metas provides any additional, non-factor information about
a given element. This makes adding information points *much* easier as
snapshot_text required that the element do all of the formatting.

Message raising is now supported by checking for the "trap_message" attribute
on a transaction's controller and *all* of its primary factors. Like
with-statement's exit, a message will not raised if a True value is returned by
a given "trap_message" callable. If a message is not suffocated, it will
eventually be given to `postgresql.sys.msghook`, which can be overridden by the
user.

Warnings are no longer Python Warnings, just Message sub-classes. Using Python
warnings was pointless.

Connection failures are not longer published on CCC errors. They are now
published on the connection object itself using
postgresql.protocol.client3.ConnectionAttempt objects.
This object is temporary, so it will be undocumented like CCC's
connection_failure attribute. It may not even make it into 0.9.