Relocate protocol I/O and buffer management code into a low-level client.
The existing client code combined both interface and protocol-state
management code on a single object. This was inappropriate because it required
specialized reset methods in order to accommodate for connectors that produced
multiple socket factories.

By delegating protocol-state managament to another object this allows for cleaner
connection creation code. This also provides:

 * Formalized how fatal client errors are identified and propagated.
 * Provided concrete lifetimes to connection objects.
 * Paves the way for a PQv3 abstaction. (ctypes+libpq could become a possibility)

Some additional enhancements based on this work will be:

 * All client errors will be specified via their code, and
 * TypeIO will provide code -> exception/message/warning mapping.