Refactor the implementation of Cursor and Chunks.
The prior implementation was based on the previous API design where a statement
could "only" be executed to have a cursor returned. In order to get 0.8 out the
door, the effect of that design remained.

The current API provides multiple execution methods to match the user's needs.
In turn, Cursors no longer have local buffer state, and Chunks have extremely
simple state. Stream(Chunks)-vs-Random Access(Cursor).

This removes the chunksize keyword argument. If the size needs to be configured,
the user will need to instantiate the chunk and modify the property on the Chunk
object. chunksize has no effect on Cursor objects.