Add protocol optimizations; pack_tuple_data and cat_messages.
 * Provides a more efficient means to construct the send buffer from the out-going
   messsages.
 * pack_tuple_data is used by Tuple.serialize to quickly render the serialized
   form.
 * add ulong and ushort optimizations.
 * minor refactoring to parse_tuple_message.

Notably, the new cat_messages function will take note of bytes() objects and
treat them as COPY data. This optimization allows the avoidance of creating
CopyData messages for each COPY line.

Arguably, at the client3 level, this optimization is inappropriate. However, it
is unlikely for there to be another case where this implicit message typing is
desirable for PQv3. It is likely that any client3 extensions would take on
formal element3 subclasses, and therefore not require the exercise of such
functionality. Additionally, if implicit typing were to be desireable, it would
require additional configuration of some sort even in the absence of this
optimization. Perhaps the parameterization of the implied message type would be
appropriate compensation for advanced applications..