Fixes stored procedure invocation statement generation.
Bug 1010620 reported by Dallas Morisette.

The former implementation depended on regtype's string represention matching
the type identifiers. This was not the case for varchar as it was rendered as
'character varying'. Statement generation could have been corrected while using
that method, but it would have meant including exception cases. So, in order to
encourage simplicity here, use fully-qualified type identifiers to reference
the parameter types.