Generalize exception instantiation in pq3.
This is being done with the purpose of allowing the error class to be defined by
the driver's selected TypeIO. Currently, it's directly specified by the
Connection as there are still places that rely on postgresql.exceptions being
raised.

Also, use ClientError's POSITION to indicate which attribute failed to pack or
unpack in Type I/O error.