Implement cluster control for win32.
This allows the tests to be executed on MSW operating systems. Some test
exceptions had to be made for test_connect as IPv6 is not turned on by default
on the windows builds, and there doesn't appear to be a consistent way to detect
that it's on.

- Add minor tests for postgresql.installation.
- Add a python.os module for providing routines for finding executables
  and properly extending exe names on win32.
- Remove the _e_factors from Installation objects as they do not necessarily
  come from pg_config binaries or files.

Some cleanup work is still needed,
so minor refactoring will occur around this area.

closes #4