Add some overflow checks for cat_messages.
Some additional work is still necessary in this area, but this should cover the
concerned corners. More tests would be nice.

Some attention should be paid to Py_ssize_t size for size_t size, and often the
casts to uint32_t may not be appropriate. However, these would be substantially
strange circumstances considering that such a thing would mean that a gigabyte
sized bytes() object was passed in.

Fixes #26