Use the buffer API.
The direct use of bytes, while desirable for speed, limits the number of types the
unpack routines can work with. Eventually, we may be using memoryviews more
often, so it's imperative to support the buffer APIs.
[Didn't see much, if any, slow down on the single int test either.]

Also, relax the check on the buffer's size to only require that *at least* 2,
4, or 8 bytes exist in the buffer. This further aligns the implementation with
Struct.unpack_from.