Implement MessageBuffer.getvalue()
Like io.BytesIO, this gets the entire contents of the buffer as a bytes object.
The difference is that the initial offset is respected; any messages read out of
the buffer don't count as being part of the value regardless of the data still
existing in the buffer.

This is a requisite for Zero-COPY.