Implement consume_tuple_messages.
This offers a C optimization for processing Tuple('D') messages in
put_tupledata. The method will continue parsing messages until the message type
is not 'D'. The existing pure-Python implementation proved to be rather costly
as the invocation overhead of the comparison and parse_tuple_messages are
frequent. consume_tuple_messages also offered a more appropriate place to
implement a 'takewhile' iterator, which allows put_tupledata to return
the number consumed. This means that put_tupledata can now be more optimistic
and avoid double processing of tuple data near command completion.