Add tests for reconciliation.
s/reconciliation/realignment/ in most cases. The terminology is now realignment
deals with connection state, and reconciliation deals with recovering from
[receiver] faults.

Change load_chunks to use a SynchronizeMessage instead of a Flush; apparently
in COPY failures, a sync is performed regardless, so xact3.Instruction won't be
expecting that. Arguably, using a sync here is preferrable, but Flush had some
nice properties when working outside of a transaction block. If one failed, they
all failed. This was necessary in order to handle realignment in COPY fail
cases.