Remove postgresql.unittest.
test_connect is the only module that should be using postgresql.unittest, so
remove to make sure people don't get any ideas about how it's intended to be
used.

postgresql.temporal should be used in most cases and its lack of documentation
should be a good indicator that using it is currently not appropriate.