Fix automatic cleanup of cursors and statements
Current coding for automatic collection and closure of cursors and
statements is using a callback through a weakref, but stores this
weakref _inside_ the tracked object, thus making the weakref part of a
potential reference cycle.  Per Python spec, the callback will only be
invoked if the weakref is still alive, which will not be the case if a
cursor or a statement were part of a reference cycle.  This leads to a
failure to close the cursor on database side and a subsequent
DuplicateCursorError if the id got reused for another cursor.

This is easily fixed by storing the weakref alongside the cursor and
statement garbage lists.